--- src/app/(route)/list/_components/PostWriteMenu/WRITE_MENU_STYLES.ts
@@ -1,5 +1,0 @@
-export const WRITE_MENU_STYLES = {
-  menuButton: "flex items-center gap-2 px-7 py-4",
-  menuLabel: "text-h3-medium text-neutral-normal-default",
-  findModalButton: "min-h-11 w-full",
-};
