--- src/app/(route)/list/_constants/SEARCH_HEADER_TITLE.ts
@@ -1,5 +1,0 @@
-export const SEARCH_HEADER_TITLE = {
-  default: "게시글",
-  region: "지역 선택",
-  post: "게시글 검색",
-};
