--- src/app/(route)/list/_constants/TABS.ts
@@ -1,10 +1,0 @@
-export const TABS = [
-  {
-    key: "lost",
-    label: "분실",
-  },
-  {
-    key: "found",
-    label: "발견",
-  },
-];
